Added the original code file
The code for the program (Operation ThunderSnatch).

The options for how to calculate things are displayed out for the user. 

Resulting calculation is a little unsightly, may come back and tweak the data items to change how many numbers are displayed as well as where the decimal point is located.